+++ src/features/webSearch.ts
@@ -2,6 +2,7 @@ import * as vscode from 'vscode'
 import * as https from 'https'
 import * as tmp from 'tmp'
 import * as fs from 'fs'
+import { outputChannel } from '../outputChannel'
 
 /**
  * Performs a web search using the Jina AI search engine and displays the results in a Cody AI mention.
@@ -21,6 +22,7 @@ export async function webSearch(): Promise<void> {
   if (!extension) {
     // Show a warning to the user that the extension is not active or installed
     vscode.window.showWarningMessage('Cody AI extension is not active or installed')
+    outputChannel.appendLine('WebSearch: Cody AI extension is not active or installed')
     return
   }
 
@@ -36,9 +38,12 @@ export async function webSearch(): Promise<void> {
         return
       }
 
+      outputChannel.appendLine(`WebSearch: Gathering the web result for "${query}"`)
       // Encode the query
       const encodedQuery = encodeURIComponent(query)
       const url = `https://s.jina.ai/${encodedQuery}`
+
+      outputChannel.appendLine(`WebSearch: Gathering the web result at "${url}"`)
 
       // Create a status bar item for the progress indicator
       const statusBarItem = vscode.window.createStatusBarItem(vscode.StatusBarAlignment.Right)
@@ -69,7 +74,16 @@ export async function webSearch(): Promise<void> {
 
         // Handle the response data
         response.on('data', chunk => {
+          outputChannel.appendLine('WebSearch: Recieving chunk: ' + chunk)
           data += chunk
+        })
+
+        response.on('error', (err: any) => {
+          // Clear the progress interval and hide the status bar item
+          outputChannel.appendLine('WebSearch: Error with code: ' + err)
+          clearInterval(progressInterval)
+          statusBarItem.hide()
+          statusBarItem.dispose()
         })
 
         // Handle the response end
@@ -92,6 +106,7 @@ export async function webSearch(): Promise<void> {
 
         // Show an error message to the user
         vscode.window.showErrorMessage('An error occurred while making the HTTP request.')
+        outputChannel.appendLine('WebSearch: An error occurred while making the HTTP request.')
       })
     })
 }
@@ -109,14 +124,25 @@ export async function displaySearchResultsInMention(query: string, message: stri
     const workspaceFolders = vscode.workspace.workspaceFolders
     if (workspaceFolders) {
       const workspaceFolder = workspaceFolders[0]
-      fs.mkdirSync(workspaceFolder.uri.fsPath + '/.codyarchitect/webresults', { recursive: true })
+      const path = fs.mkdirSync(workspaceFolder.uri.fsPath + '/.codyarchitect/webresults', {
+        recursive: true
+      })
+      if (path) {
+        outputChannel.appendLine(
+          'WebSearch: displaySearchResultsInMention: Path for Web search results created at ' + path
+        )
+      }
       const file = vscode.Uri.file(
         workspaceFolder.uri.fsPath + '/.codyarchitect/webresults/' + query + '.md'
       )
       fs.writeFileSync(file.fsPath, Buffer.from(content))
       await vscode.commands.executeCommand('cody.mention.file', file)
+      outputChannel.appendLine(
+        'WebSearch: displaySearchResultsInMention: Web search results created'
+      )
     }
-  } catch (err) {
+  } catch (err: any) {
     console.error(err)
+    outputChannel.appendLine('WebSearch: displaySearchResultsInMention: ' + err)
   }
 }