+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 ## Change-Log:
 
+### Version 0.0.12 (2024-06-19)
+
+- Add logging to readPDF and webSearch features
+
 ### Version 0.0.11 (2024-06-18)
 
-- Web Results and PDF results will be saved in the `codyarchitect/webresult`, `codyarchitect/pdfresult` folders respectively.
+- Web Results and PDF results will be saved in the `codyarchitect/webresults`, `codyarchitect/pdfresults` folders respectively.